+++ src/WtCore/HftStraContext.h
@@ -33,11 +33,11 @@ public:
 	
 	virtual void on_schedule(uint32_t uDate, uint32_t uTime) override;
 
-	virtual void on_trade(const char* stdCode, bool isBuy, uint32_t vol, double price) override;
+	virtual void on_trade(const char* stdCode, bool isBuy, double vol, double price) override;
 
-	virtual void on_order(uint32_t localid, const char* stdCode, bool isBuy, uint32_t totalQty, uint32_t leftQty, double price, bool isCanceled = false) override;
+	virtual void on_order(uint32_t localid, const char* stdCode, bool isBuy, double totalQty, double leftQty, double price, bool isCanceled = false) override;
 
-	virtual void on_position(const char* stdCode, bool isLong, uint32_t prevol, uint32_t preavail, uint32_t newvol, uint32_t newavail) override;
+	virtual void on_position(const char* stdCode, bool isLong, double prevol, double preavail, double newvol, double newavail) override;
 
 	virtual void on_channel_ready() override;
